Optimized AP onClear stuff...
Optimized onClear stuff to minimize redundant changes to obj.Active of JsonItems, because that was super slow.
When onClear is called, it'll set up a temporary table to keep track of the soon-to-be JsonItem Active states, wait for about 64 frames to make sure all of the relevant data is received from the AP server, *then* update the Active state of each JsonItem. Instead of setting the Active state of each JsonItem to false, then one by one setting them to true as-needed.
If syncing with the AP server from a refreshed/cleared state, the new code will basically be the same as before performance-wise. But if syncing with the AP server from one's most recent autosave, this can save a lot of time processing.